+++ packages/pn-personagiuridica-webapp/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.22.2](https://fbianchicodermine-github/pagopa/pn-frontend/compare/v2.22.2-RC.0...v2.22.2) (2026-04-13)
+
+**Note:** Version bump only for package @pagopa-pn/pn-personagiuridica-webapp
+
+
+
+
+
+## [2.22.2-RC.0](https://fbianchicodermine-github/pagopa/pn-frontend/compare/v2.22.1...v2.22.2-RC.0) (2026-04-13)
+
+**Note:** Version bump only for package @pagopa-pn/pn-personagiuridica-webapp
+
+
+
+
+
+## [2.22.1](https://github.com/pagopa/pn-frontend/compare/v2.22.1-RC.0...v2.22.1) (2026-04-08)
+
+**Note:** Version bump only for package @pagopa-pn/pn-personagiuridica-webapp
+
+
+
+
+
 # [2.22.0](https://github.com/pagopa/pn-frontend/compare/v2.22.0-RC.4...v2.22.0) (2026-03-23)
 
 **Note:** Version bump only for package @pagopa-pn/pn-personagiuridica-webapp